requirements.txt
# Task ID: T003 - Backend dependencies
# Core framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlmodel>=0.0.14
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.25
alembic>=1.13.0

# Authentication
python-jose[cryptography]>=3.3.0

# Environment
python-dotenv>=1.0.0

# CORS
# (included in fastapi)

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
